refactor(markdown): share code block chrome
