+++ src/components/markdown/ProjectedMarkdownView.tsx
@@ -1,7 +1,6 @@
-import { Check, Copy } from "lucide-react";
-import { Fragment, useState, type CSSProperties, type ReactNode } from "react";
+import { Check } from "lucide-react";
+import { Fragment, type CSSProperties, type ReactNode } from "react";
 import katex from "katex";
-import { StaticCodeBlock } from "@/components/editor/static-highlight";
 import type {
   MarkdownProjectionBlock,
   MarkdownProjectionPlan,
@@ -12,13 +11,9 @@ import { useColorTheme, useDarkMode } from "@/lib/dark-mode";
 import { katexStrict } from "@/lib/katex-options";
 import { cn } from "@/lib/utils";
 import type { MarkdownHeadingAnchor } from "@/components/outputs/markdown-heading-anchors";
+import { MarkdownCodeBlock } from "./MarkdownCodeBlock";
 import {
   markdownBlockquoteClassName,
-  markdownCodeBlockCopyButtonClassName,
-  markdownCodeBlockLabelClassName,
-  markdownCodeBlockPreStyle,
-  markdownCodeBlockShellClassName,
-  markdownCodeBlockToolbarClassName,
   markdownDeleteClassName,
   markdownDisplayMathClassName,
   markdownDocumentClassName,
@@ -877,59 +872,15 @@ function ProjectedCodeBlock({
   isDark: boolean;
   language?: string;
 }) {
-  const [copied, setCopied] = useState(false);
-
-  const copyCode = async () => {
-    try {
-      await navigator.clipboard.writeText(code);
-      setCopied(true);
-      setTimeout(() => setCopied(false), 1800);
-    } catch (error) {
-      console.error("Failed to copy projected markdown code block:", error);
-    }
-  };
-
-  const languageLabel = codeBlockLanguageLabel(language);
-
   return (
-    <div
-      data-slot="markdown-code-block"
-      className={markdownCodeBlockShellClassName}
-      data-code-language={languageLabel === "code" ? undefined : languageLabel}
-    >
-      <div className={markdownCodeBlockToolbarClassName}>
-        <span
-          className={markdownCodeBlockLabelClassName}
-          title={languageLabel === "code" ? "Code block" : `${languageLabel} code block`}
-        >
-          {languageLabel}
-        </span>
-        <button
-          type="button"
-          aria-label={copied ? "Copied code" : "Copy code"}
-          className={markdownCodeBlockCopyButtonClassName}
-          title={copied ? "Copied" : "Copy code"}
-          onClick={copyCode}
-        >
-          {copied ? <Check className="size-3" /> : <Copy className="size-3" />}
-        </button>
-      </div>
-      <StaticCodeBlock
-        code={code}
-        colorTheme={colorTheme}
-        isDark={isDark}
-        language={language}
-        className="max-w-full"
-        style={markdownCodeBlockPreStyle}
-      />
-    </div>
+    <MarkdownCodeBlock
+      code={code}
+      colorTheme={colorTheme}
+      isDark={isDark}
+      language={language}
+      preClassName="max-w-full"
+    />
   );
-}
-
-function codeBlockLanguageLabel(language: string | undefined): string {
-  const trimmed = language?.trim();
-  if (!trimmed) return "code";
-  return trimmed;
 }
 
 function ProjectedMath({ displayMode = false, latex }: { displayMode?: boolean; latex: string }) {
